refact: creating common module to have middlewares, logger and database creation functions in one single place
