+++ api-gateway/go.mod
@@ -21,3 +21,5 @@ require (
 	golang.org/x/tools v0.1.12 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
+
+replace github.com/FelipeFelipeRenan/goverse/common => ../common
