+++ room-service/go.mod
@@ -37,3 +37,5 @@ require (
 )
 
 replace github.com/FelipeFelipeRenan/goverse/proto => ../proto
+
+replace github.com/FelipeFelipeRenan/goverse/common => ../common
