+++ user-service/go.mod
@@ -4,16 +4,20 @@ go 1.24.2
 
 require (
 	github.com/FelipeFelipeRenan/goverse/proto v0.0.0-20250706201258-4a18dd1e3946
-	github.com/jackc/pgx/v5 v5.7.5
+	github.com/jackc/pgx/v5 v5.7.6
 	github.com/joho/godotenv v1.5.1
-	github.com/stretchr/testify v1.10.0
-	golang.org/x/crypto v0.39.0
+	github.com/stretchr/testify v1.11.1
+	golang.org/x/crypto v0.41.0
 	google.golang.org/grpc v1.73.0
 )
 
-require github.com/jackc/puddle/v2 v2.2.2 // indirect
+require (
+	github.com/jackc/puddle/v2 v2.2.2 // indirect
+	go.yaml.in/yaml/v2 v2.4.2 // indirect
+)
 
 require (
+	github.com/FelipeFelipeRenan/goverse/common v0.0.0-00010101000000-000000000000
 	github.com/beorn7/perks v1.0.1 // indirect
 	github.com/cespare/xxhash/v2 v2.3.0 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
@@ -22,19 +26,21 @@ require (
 	github.com/kr/text v0.2.0 // indirect
 	github.com/munnerz/goautoneg v0.0.0-20191010083416-a7dc8b61c822 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/prometheus/client_golang v1.22.0
+	github.com/prometheus/client_golang v1.23.2
 	github.com/prometheus/client_model v0.6.2 // indirect
-	github.com/prometheus/common v0.65.0 // indirect
+	github.com/prometheus/common v0.66.1 // indirect
 	github.com/prometheus/procfs v0.17.0 // indirect
 	github.com/rogpeppe/go-internal v1.14.1 // indirect
 	github.com/stretchr/objx v0.5.2 // indirect
-	golang.org/x/net v0.41.0 // indirect
+	golang.org/x/net v0.43.0 // indirect
 	golang.org/x/sync v0.16.0 // indirect
-	golang.org/x/sys v0.34.0 // indirect
-	golang.org/x/text v0.27.0 // indirect
+	golang.org/x/sys v0.35.0 // indirect
+	golang.org/x/text v0.28.0 // indirect
 	google.golang.org/genproto/googleapis/rpc v0.0.0-20250707201910-8d1bb00bc6a7 // indirect
-	google.golang.org/protobuf v1.36.6 // indirect
+	google.golang.org/protobuf v1.36.8 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
 replace github.com/FelipeFelipeRenan/goverse/proto => ../proto
+
+replace github.com/FelipeFelipeRenan/goverse/common => ../common
