--- user-service/middleware/chain_middleware.go
@@ -1,14 +1,0 @@
-package middleware
-
-import "net/http"
-
-type Middleware func(http.HandlerFunc) http.HandlerFunc
-
-func ChainMiddleware(middlewares ...Middleware) Middleware {
-	return func(final http.HandlerFunc) http.HandlerFunc {
-		for i := len(middlewares) - 1; i >= 0; i-- {
-			final = middlewares[i](final)
-		}
-		return final
-	}
-}
