--- user-service/pkg/database/pgx.go
@@ -1,93 +1,0 @@
-package database
-
-import (
-	"context"
-	"fmt"
-	"os"
-	"time"
-
-	"github.com/FelipeFelipeRenan/goverse/user-service/pkg/logger"
-	"github.com/jackc/pgx/v5/pgxpool"
-	"github.com/joho/godotenv"
-)
-
-func LoadEnv() error {
-	err := godotenv.Load(".env")
-	if err != nil {
-		return fmt.Errorf("Erro ao carregar configurações do banco de dados: %v", err)
-	}
-	return nil
-}
-
-func Connect() (*pgxpool.Pool, error) {
-	// if os.Getenv("ENV") != "prod" {
-	// 	if err := godotenv.Load(); err != nil {
-	// 		logger.Error("Erro ao carregar .env", "err", err)
-	// 	}
-	// }
-
-	dbHost := os.Getenv("DB_HOST")
-	dbPort := os.Getenv("DB_PORT")
-	dbUser := os.Getenv("DB_USER")
-	dbPassword := os.Getenv("DB_PASSWORD")
-	dbName := os.Getenv("DB_NAME")
-
-	connStr := fmt.Sprintf("postgres://%s:%s@%s:%s/%s?sslmode=disable", dbUser, dbPassword, dbHost, dbPort, dbName)
-
-	var pool *pgxpool.Pool
-	var err error
-
-	maxRetries := 5
-	for i := 0; i < maxRetries; i++ {
-		pool, err = pgxpool.New(context.Background(), connStr)
-		if err == nil {
-			// Tenta fazer um ping para garantir que a conexão está realmente viva
-			if pingErr := pool.Ping(context.Background()); pingErr == nil {
-				logger.Info("Conexão com o banco de dados estabelecida com sucesso")
-				return pool, nil
-			} else {
-				err = pingErr // Guarda o erro do ping para o log
-			}
-		}
-
-		logger.Warn("Não foi possível conectar ao banco de dados. Tentando novamente...",
-			"tentativa", i+1,
-			"max_tentativas", maxRetries,
-			"erro", err.Error(),
-		)
-		time.Sleep(5 * time.Second) // Espera 5 segundos antes de tentar de novo
-	}
-
-	return nil, fmt.Errorf("não foi possível conectar ao banco de dados após %d tentativas: %w", maxRetries, err)
-}
-
-func RunMigration(pool *pgxpool.Pool) error {
-	ctx := context.Background()
-
-	_, err := pool.Exec(ctx, `CREATE EXTENSION IF NOT EXISTS "uuid-ossp";`)
-	if err != nil {
-		return fmt.Errorf("erro ao ativar extensão uuid-ossp: %w", err)
-	}
-
-	query := `
-		CREATE TABLE IF NOT EXISTS users (
-			id UUID PRIMARY KEY DEFAULT uuid_generate_v4(),
-			username TEXT NOT NULL,
-			email TEXT UNIQUE NOT NULL,
-			password TEXT NOT NULL,
-			picture TEXT,
-			created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-			is_oauth BOOLEAN,
-			updated_at TIMESTAMP,
-			deleted_at TIMESTAMP NULL
-	);	
-	`
-
-	_, err = pool.Exec(ctx, query)
-	if err != nil {
-		return fmt.Errorf("erro ao executar migração: %w", err)
-	}
-
-	logger.Info("Migração executada com sucesso...")
-	return nil
-}
